create netlink-packet-utils crate
There is a lot of code that is redundant in all the
`netlink-packet-xxx` crates:

- the Parseable/ParseableParametrized/Emitable traits
- the Nla types: Nla, NlaBuffer, NlaIterator
- the error types
- the buffer! macro

Instead of copy-pasting this code, we gathered it in this
netlink-packet-utils crate.

Note that due to the orphan rule, having the Parseable/Emitable traits
in a separate crate may cause problems because we might not be able to
implement them for certain types. To mitigate this issue, we changed
the Parseable and ParseableParametrized:

```
pub trait Parseable<T> {
    fn parse(&self) -> Result<T, DecodeError>;
}
```

becomes:

```
pub trait Parseable<T>
where
    Self: Sized,
{
    fn parse(buf: &T) -> Result<Self, DecodeError>;
}
```